adds Services, DTOs, etc...

+++ src/main/java/com/miguelcordoba/admservice/persistence/repository/DocumentRepository.java
@@ -3,5 +3,5 @@ package com.miguelcordoba.admservice.persistence.repository;
 import com.miguelcordoba.admservice.persistence.entity.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentRepository extends JpaRepository<Document, Integer> {
+public interface DocumentRepository extends JpaRepository<Document, Long> {
 }
